backend/flask/requirements.txt
Flask>=3.0,<4
Flask-SQLAlchemy>=3.1,<4
Flask-JWT-Extended>=4.6,<5
Flask-Cors>=4.0,<5
marshmallow>=3.21,<4
python-dotenv>=1.0,<2
Werkzeug>=3.0,<4
SQLAlchemy>=2.0,<3
PyMySQL>=1.1,<2
